Added wrappers to inlineOneButtonSelect
